#86 feat: 탐험 시작 시 다이나믹 아일랜드 장소명 반영 구현
